# api/openapi.yaml
openapi: 3.1.0
info:
  title: ShowMesh Control API
  version: "1"
  summary: >-
    The coordinator's public, versioned control API and Server-Sent
    Events change stream (ADR-014).
  description: >-
    This is the versioned public contract ADR-014 requires: it is designed
    to be used without any Operator UI, by a CLI (`showmeshctl`), an
    automation system, or an alternate client. Everything here is served
    under `/api/v1`; the container healthcheck endpoints `/healthz`,
    `/readyz`, and `/version` are infrastructure probes outside this
    versioned contract and are not documented here.


    **Versioning is additive-only within v1.** Fields may be added to any
    response; they are never removed, renamed, or retyped. A client MUST
    ignore fields it does not recognize. A breaking change is a `/api/v2`
    contract, not an edit to this one.


    **`serverTime` appears on every response, including every error.** A
    client computes evidence ages against `serverTime` rather than its own
    clock, so clock skew between the client and the coordinator is visible
    instead of silently wrong. Payloads never carry a precomputed age or
    "secondsAgo" field for exactly this reason: an age field would change on
    every poll even when nothing else did, which would make change
    detection on the SSE stream fire continuously for no real change.


    **`offline` means the control-plane (MQTT) connection is gone, not that
...


...
    `state: "unknown_age"`.


...


...


...


...

...
servers:
...
security:
  - bearerAuth: []
  - cookieAuth: []
  - {}
paths:
...
  /observations:
    get:
      operationId: listObservations
      summary: Flat observation list
      description: >-
        Every observation this coordinator currently holds, across every
        resource, optionally filtered. A filter that matches nothing
        returns `200` with an empty `observations` array — "no observations
        match this filter" is not "this endpoint does not exist", so this
        is never a `404`.
      parameters:
        - name: resourceKind
          in: query
          required: false
          schema:
            type: string
            enum: [node, fpp, coordinator]
          description: Restrict to observations about resources of this kind.
        - name: resourceId
          in: query
          required: false
          schema:
            type: string
          description: Restrict to observations about this specific resource ID.
        - name: signal
          in: query
          required: false
          schema:
            type: string
          description: Restrict to observations of this exact signal ID.
      responses:
        "200":
          description: OK
          headers:
            ShowMesh-API-Version:
              $ref: "#/components/headers/ShowMesh-API-Version"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObservationsResponse"
        "400":
          $ref: "#/components/responses/InvalidParameter"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalError"
  /events:
    get:
      operationId: listEvents
      summary: Event history
      description: >-
        Event history in ascending `seq` order. See this document's
        top-level description for what `gap: true` means; it is a
        successful response describing an incomplete answer, never an
        error.
      parameters:
        - name: since
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: >-
            Exclusive lower bound: return events with seq strictly greater
            than this value. 0 (the default) means "from the beginning of
            retained history".
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 100
            maximum: 500
          description: >-
            Maximum number of events to return. Defaults to 100; values
            above 500 are silently clamped to 500. Both numbers are this
            coordinator's own pagination choice, not a measured or
            contractually significant value (see RES-013), and may change.
      responses:
        "200":
          description: OK
          headers:
            ShowMesh-API-Version:
              $ref: "#/components/headers/ShowMesh-API-Version"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventsResponse"
        "400":
          $ref: "#/components/responses/InvalidParameter"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  responses:
...
    InvalidParameter:
      description: A path or query parameter failed validation.
      headers:
        ShowMesh-API-Version:
          $ref: "#/components/headers/ShowMesh-API-Version"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: >-
        No valid credential was presented: reads are closed and none was
        given (or it did not authenticate), or the route is a write /
        `GET /audit` and no credential authenticated at all.
      headers:
        ShowMesh-API-Version:
          $ref: "#/components/headers/ShowMesh-API-Version"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: >-
        A valid credential authenticated, but the principal does not hold
        the scope this operation requires. `detail` names the missing
        scope (ADR-024 decision 4).
      headers:
        ShowMesh-API-Version:
          $ref: "#/components/headers/ShowMesh-API-Version"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    MethodNotAllowed:
      description: >-
        The path names a real route, but this request's method is not one
        that route serves. A method this document does not define for a
        path returns this response rather than a `404`, and the `Allow`
        header names exactly the methods that path does serve.

        This description used to say that the document defined only `get`
        operations because every route was read-only. That stopped being
        true in Step 6, which added `POST /session` and
        `POST /bootstrap`, and stopped being true three times over in
        Step 7, which added the configuration, discovery and FPP command
        writes. What has NOT changed is the rule underneath it: no state
        change in this API is reachable by `GET` (ADR-024), so a `GET`
        remains safe to issue against any path here.
      headers:
        ShowMesh-API-Version:
          $ref: "#/components/headers/ShowMesh-API-Version"
        Allow:
          $ref: "#/components/headers/Allow"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    InternalError:
      description: >-
        The coordinator failed to answer this request for a reason that
        has nothing to do with whether the requested resource exists or
        the request was well formed — an internal dependency (e.g. the
        store) failed. `detail` never carries the underlying error; it is
        logged coordinator-side only.
      headers:
        ShowMesh-API-Version:
          $ref: "#/components/headers/ShowMesh-API-Version"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  headers:
    ShowMesh-API-Version:
      description: >-
        The API major version this response was served by. Present on
        every `/api/v1` response, success or error, with no exception
        (contract section 6.2). A client MAY also send this as a request
        header naming the version it expects; a coordinator that does not
        serve that version answers with the `UnsupportedAPIVersion`
        problem instead of a partial or best-guess render (contract
        section 6.6).
      schema:
        type: string
        enum: ["1"]
    Allow:
      description: >-
        Present only on a `405` response. The exact, comma-separated set
        of HTTP methods the requested path serves (e.g. `"GET, HEAD"`),
        computed from the routes actually registered — never a static or
        hand-maintained list.
      schema:
        type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        `Authorization: Bearer <token>` (ADR-024). token is an API token
        minted for a principal by an admin, displayed exactly once at
        creation and stored coordinator-side only as a SHA-256 digest. A
        malformed or non-`Bearer` `Authorization` header never falls
        through to `cookieAuth` (ADR-024 decision 6) — it is treated as no
        credential at all. Reads are open by default (this scheme is
        optional for them, per the top-level `security` requirement not
        being exclusive); every write requires it or `cookieAuth`, with no
        opt-out. `/healthz`, `/readyz`, and `/version` are outside this
        contract entirely and are never authenticated, at any
        configuration.
    cookieAuth:
      type: apiKey
      in: cookie
      name: showmesh_session
      description: >-
        The `HttpOnly` session cookie `POST /session` mints (ADR-024
        decision 5). Not usable outside a browser in practice (nothing
        exposes it to script), and a write authenticated this way
        additionally requires the `Sec-Fetch-Site: same-origin` request
        header (decision 6) — omitted from this scheme's own contract
        because OpenAPI's securityScheme vocabulary has no way to express
        a conditional required header; see this document's top-level
        description and `components.responses.CSRFRejected`.
  schemas:
...
    ResourceRef:
      type: object
      required: [kind, id]
      properties:
        kind:
          type: string
          enum: [node, fpp, coordinator]
        id:
          type: string
...
    ObservationEntry:
      type: object
      description: >-
        One element of GET /observations: the section 6.3 Evidence envelope
        plus its own resource, since a flat list spanning every resource
        cannot rely on an enclosing object to imply the subject the way
        Node.evidence.* or FPPInstance.observations can.
      required:
        [resource, signal, value, unit, state, reason, observedAt,
         collectedAt, source, quality, validForSeconds]
      properties:
        resource:
          $ref: "#/components/schemas/ResourceRef"
        signal:
          type: string
        value:
          type: [boolean, string, integer, number, "null"]
        unit:
          type: [string, "null"]
        state:
          type: string
          enum:
            [current, stale, unknown_age, not_collected, collection_failed,
             unsupported]
        reason:
          type: [string, "null"]
        observedAt:
          type: [string, "null"]
          format: date-time
        collectedAt:
          type: [string, "null"]
          format: date-time
          description: See Evidence.collectedAt; null only for state "not_collected".
        source:
          type: string
        quality:
          type: string
          enum: [direct, derived, inferred, operator]
        validForSeconds:
          type: [integer, "null"]
...
    ObservationsResponse:
      type: object
      required: [serverTime, observations]
      properties:
        serverTime:
          type: string
          format: date-time
        observations:
          type: array
          description: >-
            Ordered by resourceKind, then resourceId, then signal,
            ascending, and this ordering is guaranteed.
          items:
            $ref: "#/components/schemas/ObservationEntry"
    EventsResponse:
      type: object
      description: >-
        See this document's top-level description for gap and
        oldestRetainedSeq: gap is always present and boolean;
        oldestRetainedSeq is the lowest seq still retained, or null when
        the store retains no events at all.
      required: [serverTime, events, latestSeq, gap, oldestRetainedSeq]
      properties:
        serverTime:
          type: string
          format: date-time
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        latestSeq:
          type: integer
        gap:
          type: boolean
        oldestRetainedSeq:
          type: [integer, "null"]
...